requirements-dev.txt
flake8==6.1.0
pycodestyle==2.11.0
black==23.7.0

mysql-connector-python==8.0.33
